% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implementation.R
\name{myGeneric,character-method}
\alias{myGeneric,character-method}
\title{A Generic}
\usage{
\S4method{myGeneric}{character}(x, y, ...)
}
\arguments{
\item{x}{Describe x.}

\item{y}{Describe y.}

\item{...}{Describe ... .}
}
\description{
This generic always takes x,y, other methods might use ...
}
